+++ arn/tree_test.go
@@ -14,7 +14,7 @@ should all match the following topics
 
 https://docs.oasis-open.org/mqtt/mqtt/v5.0/os/mqtt-v5.0-os.html#_Toc3901241
 */
-package ftree
+package arn
 
 import (
 	"bytes"
