--- ftree/testdata/golden.files
@@ -1,1 +1,0 @@
-ftree.TestTree_Filters
